# openapi.yaml
openapi: 3.0.3
info:
  title: auto_fix_mcp sandbox API (Fastn-resembling)
  version: "2.0.0"
  description: >
    Deliberately vulnerable multi-tenant RBAC sandbox, built to resemble the
    shape of a real client engagement (2 tenants, 6-role hierarchy, systemic
    missing role/tenant checks). See VULNERABILITIES.md in the repo for the
    planted issues and how each one maps to the endpoints below. Seed
    accounts are documented there too.
servers:
  - url: https://auto-fix-mcp.onrender.com
    description: Live (Render)
  - url: http://localhost:3000
    description: Local

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Real signed JWT obtained from POST /api/v1/auth/login. Signature is
        verified, but expiry is intentionally ignored server-side (see
        "Expired Authorization" in VULNERABILITIES.md).

  schemas:
    Org:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        plan: { type: string }
        billingEmail: { type: string }

    OrgUpdateRequest:
      type: object
      description: No field allow-list is enforced (Mass Assignment) — any property is accepted.
      additionalProperties: true

    User:
      type: object
      description: Password is stored in plaintext in data.js but never serialized into this response shape (Shadow Data fix).
      properties:
        id: { type: string }
        email: { type: string }
        role:
          type: string
          enum: [owner, operator, developer, viewer, end_user, customer]
        orgId: { type: string }

    RoleUpdateRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [owner, operator, developer, viewer, end_user, customer]

    ApiKey:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        secret:
          type: string
          description: Returned in plaintext (Excessive Data Exposure).

    Connector:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string }
        secret: { type: string }

    Workflow:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        status: { type: string }

    AuditLogEntry:
      type: object
      properties:
        id: { type: string }
        actor: { type: string }
        action: { type: string }
        timestamp: { type: string, format: date-time }

    Usage:
      type: object
      properties:
        apiCalls: { type: integer }
        storageGB: { type: number }
        billingPlan: { type: string }

    HubSettings:
      type: object
      properties:
        webhookUrl: { type: string }
        notifyEmail: { type: string }

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }

    LoginResponse:
      type: object
      properties:
        token: { type: string }

    Error:
      type: object
      properties:
        error: { type: string }

paths:
  /health:
    get:
      summary: Liveness check
      operationId: getHealth
      responses:
        "200":
          description: Service is up
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }

  /api/v1/auth/login:
    post:
      summary: Authenticate and receive a signed JWT
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: >
            Login succeeded. Token is RS256-signed, 15-minute expiry, with
            iss/aud/jti claims (Weak Authorization + Missing Token
            Issuer/Audience Claim fixes).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /api/v1/auth/logout:
    post:
      summary: Revoke the caller's current token
      operationId: logout
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Token's jti added to the in-memory revocation list (Missing Token Revocation fix).
          content:
            application/json:
              schema: { type: object, properties: { status: { type: string } } }
        "401": { description: Missing/invalid token, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

  /api/v1/orgs/{orgId}:
    get:
      summary: Get an org's profile
      operationId: getOrg
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: orgId, in: path, required: true, schema: { type: string }, example: "org_1" }
      responses:
        "200":
          description: >
            Owner-only, same-org-only (FIXED — BFLA + BOLA). 403 on role or
            tenant mismatch.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Org" }
        "401": { description: Missing/invalid token, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: Wrong role or wrong tenant, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    put:
      summary: Update an org's profile
      operationId: updateOrg
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: orgId, in: path, required: true, schema: { type: string }, example: "org_1" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrgUpdateRequest" }
      responses:
        "200":
          description: >
            Owner-only, same-org-only (FIXED — BFLA + BOLA). Mass Assignment
            on the request body itself remains planted — see
            VULNERABILITIES.md.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Org" }
        "401": { description: Missing/invalid token, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: Wrong role or wrong tenant, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

  /api/v1/orgs/{orgId}/audit-log:
    get:
      summary: List an org's audit log
      operationId: getAuditLog
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: orgId, in: path, required: true, schema: { type: string }, example: "org_1" }
        - { name: limit, in: query, required: false, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: offset, in: query, required: false, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Owner/operator-only, same-org-only, paginated (FIXED — BFLA + BOLA + Pagination Missing).
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/AuditLogEntry" } }
        "401": { description: Missing/invalid token, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: Wrong role or wrong tenant, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

  /api/v1/orgs/{orgId}/api-keys:
    get:
      summary: List an org's API keys
      operationId: getApiKeys
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: orgId, in: path, required: true, schema: { type: string }, example: "org_1" }
        - { name: limit, in: query, required: false, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: offset, in: query, required: false, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Owner/operator-only, same-org-only, paginated (FIXED — BFLA + BOLA + Pagination Missing).
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/ApiKey" } }
        "401": { description: Missing/invalid token, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: Wrong role or wrong tenant, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

  /api/v1/orgs/{orgId}/users:
    get:
      summary: List an org's member users
      operationId: getOrgUsers
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: orgId, in: path, required: true, schema: { type: string }, example: "org_1" }
        - { name: limit, in: query, required: false, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: offset, in: query, required: false, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: >
            Owner/operator-only, same-org-only, paginated, password stripped
            (FIXED — BFLA + BOLA + Shadow Data + Pagination Missing).
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/User" } }
        "401": { description: Missing/invalid token, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: Wrong role or wrong tenant, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

  /api/v1/orgs/{orgId}/users/{userId}/role:
    put:
      summary: Change a member's role
      operationId: updateUserRole
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: orgId, in: path, required: true, schema: { type: string }, example: "org_1" }
        - { name: userId, in: path, required: true, schema: { type: string }, example: "u6" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoleUpdateRequest" }
      responses:
        "200":
          description: >
            Owner-only, same-org-only, change is appended to the org's audit
            log (FIXED — BFLA + Privilege Escalation + Business Audit Data
            Tampering + Cost Evasion).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { description: Invalid role value, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "401": { description: Missing/invalid token, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: Wrong role or wrong tenant, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

  /api/v1/orgs/{orgId}/connectors:
    get:
      summary: List an org's connectors
      operationId: getConnectors
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: orgId, in: path, required: true, schema: { type: string }, example: "org_1" }
        - { name: limit, in: query, required: false, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: offset, in: query, required: false, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Owner/operator/developer-only, same-org-only, paginated (FIXED — BFLA + BOLA + Pagination Missing).
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Connector" } }
        "401": { description: Missing/invalid token, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: Wrong role or wrong tenant, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

  /api/v1/orgs/{orgId}/workflows:
    get:
      summary: List an org's workflows
      operationId: getWorkflows
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: orgId, in: path, required: true, schema: { type: string }, example: "org_1" }
        - { name: limit, in: query, required: false, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: offset, in: query, required: false, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Owner/operator/developer-only, same-org-only, paginated (FIXED — BFLA + BOLA + Pagination Missing).
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Workflow" } }
        "401": { description: Missing/invalid token, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: Wrong role or wrong tenant, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

  /api/v1/orgs/{orgId}/workflows/{workflowId}/deploy:
    post:
      summary: Deploy a workflow
      operationId: deployWorkflow
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: orgId, in: path, required: true, schema: { type: string }, example: "org_1" }
        - { name: workflowId, in: path, required: true, schema: { type: string }, example: "wf_1" }
      responses:
        "200":
          description: Owner/operator/developer-only, same-org-only (FIXED — BFLA + BOLA).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workflow" }
        "401": { description: Missing/invalid token, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: Wrong role or wrong tenant, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

  /api/v1/orgs/{orgId}/usage:
    get:
      summary: Get an org's usage/billing stats
      operationId: getUsage
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: orgId, in: path, required: true, schema: { type: string }, example: "org_1" }
      responses:
        "200":
          description: >
            Owner-only, same-org-only (FIXED — was the canonical pure-BOLA
            demo: same role, wrong tenant).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Usage" }
        "401": { description: Missing/invalid token, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: Wrong role or wrong tenant, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

  /api/v1/orgs/{orgId}/hub-settings:
    get:
      summary: Get an org's hub settings
      operationId: getHubSettings
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: orgId, in: path, required: true, schema: { type: string }, example: "org_1" }
      responses:
        "200":
          description: Owner/operator-only, same-org-only (FIXED — BFLA + BOLA).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HubSettings" }
        "401": { description: Missing/invalid token, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: Wrong role or wrong tenant, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }